# swagger.yaml
basePath: /
consumes:
- application/json
definitions:
  Product:
    description: Product defines the structure of the API details
    properties:
      description:
        type: string
        x-go-name: Description
      id:
        description: the id for this product
        format: int64
        minimum: 1
        type: integer
        x-go-name: ID
      name:
        type: string
        x-go-name: Name
      price:
        format: float
        type: number
        x-go-name: Price
      sku:
        type: string
        x-go-name: SKU
    required:
    - id
    type: object
    x-go-package: github.com/alwesleys/product-api/data
host: github.com/alwesleys
info:
  description: Documentation for Product API
  title: Product API
  version: 1.0.0
paths:
  /products:
    get:
      description: returns a list of products
      operationId: listProducts
      responses:
        "200":
          $ref: '#/responses/productsResponseWrapper'
      tags:
      - products
  /products/{id}:
    delete:
      description: returns a list of products
      operationId: deleteProduct
      parameters:
      - description: The id of the product to get/update/delete from DB
        format: int64
        in: path
        name: id
        required: true
        type: integer
        x-go-name: ID
      responses:
        "201":
          $ref: '#/responses/noContent'
      tags:
      - products
    get:
      description: returns a specific product
      operationId: getProduct
      parameters:
      - description: The id of the product to get/update/delete from DB
        format: int64
        in: path
        name: id
        required: true
        type: integer
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/productResponseWrapper'
      tags:
      - products
produces:
- application/json
responses:
  noContent:
    description: ""
  productResponseWrapper:
    description: A specific product returned in a response
    schema:
      $ref: '#/definitions/Product'
  productsResponseWrapper:
    description: A list of products returned in a response
    schema:
      items:
        $ref: '#/definitions/Product'
      type: array
schemes:
- http
securityDefinitions:
  basic:
    type: basic
swagger: "2.0"